need mon.h; fix signed vs unsigned

+++ job.c
@@ -1,9 +1,11 @@
 /* $Id$ */
 
+#include "mon.h"
+
 struct job *
 job_findbyid(int id)
 {
-	int n;
+	size_t n;
 	
 	for (n = 0; n < job_list.ol_cur; n++)
 		if (job_list.ol_jobs[n]->j_id == id)
